+++ lib/containers/hyper.js
