+++ lib/index.js
@@ -13,7 +13,7 @@ import * as updaterActions from './actions/updater';
 import * as sessionActions from './actions/sessions';
 import * as termGroupActions from './actions/term-groups';
 import {fetchNotifications} from './actions/notifications';
-import HyperTermContainer from './containers/hyperterm';
+import HyperTermContainer from './containers/hyper';
 import {loadConfig, reloadConfig} from './actions/config';
 import configureStore from './store/configure-store';
